zynqmp_pmufw: Fix violations for MISRA Rule 10_7
This patch fix following violation
misra_c_2012_rule_10_7_violation: If a composite expression is used
as one operand of an operator in which the usual arithmetic conversions
are performed then the other operand shall not have wider essential type.

To solve this MISRA violation do explicit type cast as per requirement.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_gic_proxy.c
@@ -43,7 +43,7 @@ static void PmWakeEventGicProxySet(PmWakeEvent* const wake, const u32 ipiMask,
 		pmGicProxy.groups[gicWake->group].setMask &= ~gicWake->mask;
 	} else {
 		u32 addr = GIC_PROXY_BASE_ADDR +
-			   GIC_PROXY_GROUP_OFFSET(gicWake->group) +
+			   GIC_PROXY_GROUP_OFFSET((u32)gicWake->group) +
 			   GIC_PROXY_IRQ_STATUS_OFFSET;
 
 		/* Write 1 into status register to clear interrupt */
